Moved the hidden states to be a variable on the function level instead of object level. Also made them zeros instead of random numbers for the initialization.
Added a relu function to the FBRNN that appeared to be missing.

When the .data attribute is taken from a tensor, it detaches from the original and creates a new tensor that does not track the gradients by default. This was what was causing no gradients to propagate backwards to the rest of the model.